breakout a jsx parser, move the pattern parser into its own file
